+++ NEWS.md
@@ -1,7 +1,8 @@
-# BayesMallows (development versions)
+# BayesMallows (development version)
 
 * Replaced deprecated ggplot2::aes_ function with ggplot2::aes.
 * Refactoring of SMC functions (#257)
+* Improved validation and documentation of SMC post-processing functions (#262)
 
 # BayesMallows 1.2.2